+++ UIElementToolTips.cpp
@@ -50,7 +50,7 @@ void uielement_t::OnMouseMove(UINT, CPoint pt) noexcept
         _TrackingGraph->InitToolInfo(m_hWnd, _TrackingToolInfo);
 
         _LastMousePos = pt;
-        _LastBandIndex = ~0llu;
+        _LastBandIndex = ~(size_t) 0;
 
         const FLOAT ScaledX = (FLOAT) ::MulDiv((int) pt.x, USER_DEFAULT_SCREEN_DPI, (int) _DPI);
         const FLOAT ScaledY = (FLOAT) ::MulDiv((int) pt.y, USER_DEFAULT_SCREEN_DPI, (int) _DPI);
@@ -174,7 +174,7 @@ void uielement_t::DeleteTrackingToolTip() noexcept
     _ToolTipControl.TrackActivate(&_TrackingToolInfo, FALSE);
 
     _TrackingGraph = nullptr;
-    _LastBandIndex = ~0llu;
+    _LastBandIndex = ~(size_t) 0;
 }
 
 /// <summary>
